Added an AI
o Currently the solution is pretty hacky. The Controller classes should
probably be Entities so the update function is called automatically, and
I'm not too sure my networking solution is correct.
o Also, the arrows now collide with players, but it should be improved,
so you don't see the arrow behind the player before it dissapears. Also
the arrow should spawn from the center of the fireing player.