Foundation for animated block entities added
Added:
o Simple animations. Currently it is hard-coded to animate a 16x16 3
frame animation.
o A time offset texture per chunk

TODO:
o More generalized system for BE animation
o Need to store BE data in a texture (like with blocks)
o Need to store the BE ID in the vertices. (Can be stored as the texture
coordinate to this BE in the BE data texture)
o BE with animations which are not affected by u_Time
o A way to set the frame index of a BE